move reg int test to its own file
